Migrated code into Android Modules
