+++ app/domain/src/main/java/com/avsoftware/domain/model/Book.kt
@@ -1,0 +1,9 @@
+package com.avsoftware.domain.model
+
+data class Book(
+    val id: String,
+    val title: String?,
+    val author: String,
+    val coverId: Int?,
+    val status: BookStatus
+)
