+++ app/domain/src/main/java/com/avsoftware/domain/model/BookStatus.kt
@@ -1,0 +1,7 @@
+package com.avsoftware.domain.model
+
+sealed interface BookStatus {
+    data object WantToRead: BookStatus
+    data object CurrentlyReading: BookStatus
+    data object AlreadyRead: BookStatus
+}
