+++ app/domain/src/main/java/com/avsoftware/domain/usecase/GetBooksUseCase.kt
@@ -1,0 +1,8 @@
+package com.avsoftware.domain.usecase
+
+import com.avsoftware.domain.model.Book
+import io.reactivex.Single
+
+interface GetBooksUseCase {
+    operator fun invoke(): Single<List<Book>>
+}
