--- app/src/main/java/com/avsoftware/quilterdemo/data/api/OpenLibraryApiService.kt
@@ -1,18 +1,0 @@
-package com.avsoftware.quilterdemo.data.api
-
-import com.avsoftware.quilterdemo.data.api.model.ReadingLogResponse
-import io.reactivex.Single
-import retrofit2.http.GET
-
-interface OpenLibraryApiService {
-
-    @GET("people/mekBot/books/want-to-read.json")
-    fun getWantToRead(): Single<ReadingLogResponse>
-
-    @GET("people/mekBot/books/currently-reading.json")
-    fun getCurrentlyReading(): Single<ReadingLogResponse>
-
-    @GET("people/mekBot/books/already-read.json")
-    fun getAlreadyRead(): Single<ReadingLogResponse>
-
-}
