--- app/src/main/java/com/avsoftware/quilterdemo/domain/model/Book.kt
@@ -1,9 +1,0 @@
-package com.avsoftware.quilterdemo.domain.model
-
-data class Book(
-    val id: String,
-    val title: String?,
-    val author: String,
-    val coverId: Int?,
-    val status: BookStatus
-)
