--- app/src/main/java/com/avsoftware/quilterdemo/domain/model/BookStatus.kt
@@ -1,7 +1,0 @@
-package com.avsoftware.quilterdemo.domain.model
-
-sealed interface BookStatus {
-    data object WantToRead: BookStatus
-    data object CurrentlyReading: BookStatus
-    data object AlreadyRead: BookStatus
-}
