+++ app/src/main/java/com/avsoftware/quilterdemo/ui/books/BookItem.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
-import com.avsoftware.quilterdemo.domain.model.Book
+import com.avsoftware.domain.model.Book
 import com.avsoftware.quilterdemo.R
 
 @Composable
