+++ app/src/main/java/com/avsoftware/quilterdemo/ui/books/BookScreen.kt
@@ -13,7 +13,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.avsoftware.quilterdemo.domain.model.Book
+import com.avsoftware.domain.model.Book
 import com.avsoftware.quilterdemo.ui.BookState
 
 @Composable
