+++ app/src/main/java/com/avsoftware/quilterdemo/ui/books/DetailsBottomSheet.kt
@@ -13,8 +13,8 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
+import com.avsoftware.domain.model.Book
 import com.avsoftware.quilterdemo.R
-import com.avsoftware.quilterdemo.domain.model.Book
 
 @Composable
 fun DetailsBottomSheet(
