+++ build.gradle.kts
@@ -7,4 +7,5 @@ plugins {
     alias(libs.plugins.kotlin.parcelize) apply false
     alias(libs.plugins.ksp) apply false
     alias(libs.plugins.hilt.plugin) apply false
+    alias(libs.plugins.android.library) apply false
 }
